refactor: remove funcionalidades nao usadas
